% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{BatchPM}
\alias{BatchPM}
\title{Trace elements concentrations for a batch process}
\usage{
BatchPM(
  kd,
  c0,
  pm,
  cmins = matrix(),
  min.props,
  melt.arg = list(),
  dont = character(0)
)
}
\arguments{
\item{kd}{Named matrix. Partition coefficients}

\item{c0}{Named vector. The composition of the source}

\item{pm}{Scalar. Degree of partial melting (0-100)}

\item{cmins}{Matrix. Composition of minerals (not used, for consistency)}

\item{min.props}{Named vector. Mineral proportions (normalized to 1)}

\item{melt.arg}{List. Further useful arguments (mostly used for correction modules, kept here for consistency)}

\item{dont}{Character vectors. Do not calculate for these elements. Not used here, kept for consistency (mostly used in correction modules)}
}
\value{
list containing the following slots:
\itemize{
  \item \code{c0} Named vector. Composition of the source, as passed to the function
  \item \code{cL} Named vector. Composition of the liquid
  \item \code{cmins} Named matrix. Composition of all individual solid mineral phases
  \item \code{cS} Named vector. Composition of the whole solid
  \item \code{min.props} Named vector. Mineral proportions in the restite (and expressed as (0-1) of the restite)
  \item \code{FF} Real. Melt proportion, 0-100, as passed to the function
  \item \code{kd} Named matrix. The Kd table, here similar to the input (included for consistency with other modelling routines)
  \item \code{DD} Named vector. The D (bulk distribution) values.
}
}
\description{
More commonly used for batch melting, but does actually also work with crystalization.
Make sure to use the same mineral/element names in all the variables
}
