% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CationCalc.R
\name{cleanCats}
\alias{cleanCats}
\title{This function makes sure FeOt exists (it is required downstream by M)
It could also be used to do further cleanup}
\usage{
cleanCats(milli)
}
\arguments{
\item{milli}{a table containing millications. It could be a data.frame
(for instance generated by GCDmodel::millications() ), a matrix
(such as generated by GCDkit::millications() ) or a vector.}
}
\value{
a data.frame with the same values, plus FeOt.
}
\description{
This function makes sure FeOt exists (it is required downstream by M)
It could also be used to do further cleanup
}
