% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{correctMnzSat}
\alias{correctMnzSat}
\title{Correction for monazite saturation, Montel 1992 w. elements partitionning}
\usage{
correctMnzSat(
  kd,
  c0,
  pm,
  cmins,
  min.props,
  melt.arg = list(),
  dont = character(0)
)
}
\arguments{
\item{kd}{Named matrix. Partition coefficients. Must include Monazite (Mnz)}

\item{pm}{Scalar. Degree of partial melting (0-100)}

\item{cmins}{Matrix. Composition of minerals (not used, for consistency)}

\item{min.props}{Named vector. Mineral proportions (normalized to 1)}

\item{melt.arg}{List. Further useful arguments.
Includes here the melt major elements (melt.arg$mjrs),trace (melt.arg$trc), water (melt.arg$H2O), temperature in KELVIN (melt.arg$TT)}

\item{dont}{Character vector. Do not calculate for these elements. Use to avoid overptinting previous saturation calc (i.e. if you use a saturation routine /after/ this one, )}
}
\value{
list containing the following slots:
\itemize{
  \item \code{c0} Named vector. Composition of the source, as passed to the function
  \item \code{cL} Named vector. Composition of the liquid
  \item \code{cmins} Named matrix. Composition of all individual solid mineral phases
  \item \code{cS} Named vector. Composition of the whole solid
  \item \code{min.props} Named vector. Mineral proportions in the restite (and expressed as (0-1) of the restite)
  \item \code{FF} Real. Melt proportion, 0-100, as passed to the function
  \item \code{kd} Named matrix. The Kd table, here similar to the input (included for consistency with other modelling routines)
  \item \code{DD} Named vector. The D (bulk distribution) values.
  \item \code{dont} Character vector. Elements that should not be touched by further correction routines
  \item \code{tot.mass} Real. The real mass of the system (it increased since we added Zr, so it is now 100% + [Zr])
  \item \code{corr.masses} Named vector. Corrected masses for each mineral, sum = tot.mass (NOT 100)
}
}
\description{
Correction for monazite saturation, Montel 1992 w. elements partitionning
}
