% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{correctMnzSatOKZ2}
\alias{correctMnzSatOKZ2}
\title{Correction for monazite saturation, Montel 1992 w. elements partitionning}
\usage{
correctMnzSatOKZ2(
  kd,
  c0,
  pm,
  cmins,
  min.props,
  melt.arg = list(),
  dont = character(0)
)
}
\arguments{
\item{kd}{Named matrix. Partition coefficients. Must include Monazite (Mnz)}

\item{pm}{Scalar. Degree of partial melting (0-100)}

\item{cmins}{Matrix. Composition of minerals (not used, for consistency)}

\item{min.props}{Named vector. Mineral proportions (normalized to 1)}

\item{melt.arg}{List. Further useful arguments.
Includes here the melt major elements (melt.arg$mjrs),trace (melt.arg$trc), water (melt.arg$H2O), temperature in KELVIN (melt.arg$TT)}

\item{dont}{Character vector. Do not calculate for these elements. Use to avoid overptinting previous saturation calc (i.e. if you use a saturation routine /after/ this one, )}
}
\description{
Correction for monazite saturation, Montel 1992 w. elements partitionning
}
