% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{.sanitize}
\alias{.sanitize}
\title{Sanitize objects that should be one (numeric) line
typically mineral proportions or C0
Make sure it's a vector
Remove NA's
Normalize to specified value}
\usage{
.sanitize(obj, normalize.TO = 0)
}
\arguments{
\item{obj}{the one-line thing to be trated}

\item{normalize.TO}{new total, if =0 then no normalization occurs}
}
\value{
a VECTOR with no empty element, and (possibly) normalized to the desired value
}
\description{
Sanitize objects that should be one (numeric) line
typically mineral proportions or C0
Make sure it's a vector
Remove NA's
Normalize to specified value
}
