% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{DummyPM}
\alias{DummyPM}
\title{Test function}
\usage{
DummyPM(
  kd,
  c0 = NULL,
  pm = NULL,
  cmins = matrix(),
  min.props = NULL,
  melt.arg = list(),
  dont = character(0)
)
}
\description{
Test function
}
