% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{kd.table}
\alias{kd.table}
\title{Read a tab-separated file containing Kd values and convert it into a properly
formatted table}
\usage{
kd.table(file, data = WR, na.iszero = F)
}
\arguments{
\item{file}{Fully qualified (i.e. with path) name of the file to read (use
\code{setwd()} as appropriate).}

\item{data}{Matrix or data frame containing the "reference" data: only
elements presnet both in \code{data} and \code{file} will be kept.}

\item{na.iszero}{If \code{TRUE}, replace missing values by zeros. Alters the
original data, of course, but allows calculations even when some values are
missing.}
}
\value{
a data.frame containing the Kd values
}
\description{
Read a tab-separated file containing Kd values and convert it into a properly
formatted table
}
