% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZrSaturation.R
\name{M_param}
\alias{M_param}
\title{Calculate the M parameter, as used by various zircon saturation equations}
\usage{
M_param(cats)
}
\arguments{
\item{cats}{a data.frame containing the cationic proportions,
probably generated by millications()}
}
\description{
Calculate the M parameter, as used by various zircon saturation equations
}
