% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{massBal}
\alias{massBal}
\title{Check whether a model honours mass balance constrains
(for debugging purposes...)}
\usage{
massBal(model, elt, tol = 0.000001)
}
\arguments{
\item{elt}{Character. The (trace) element to check mass balance for}

\item{tol}{real. Tolerance value, how close must the two values be (round of small numerical errors)}

\item{model.}{A model, generated by any of the melting/correction functions, i.e
a list with the appropriate components}
}
\description{
Check whether a model honours mass balance constrains
(for debugging purposes...)
}
