% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CationCalc.R
\name{millications}
\alias{millications}
\title{Calculate the millication composition of a rock}
\usage{
millications(WR)
}
\arguments{
\item{WR}{A WR major element composition expressed as oxides.
It can be given in any sensible form including matrix, data.frame and (named) vector}
}
\value{
A data.frame of millication proportions
}
\description{
Calculate the millication composition of a rock
}
\details{
Millications are calculated for all "known" oxides, in GCDmodel::petroOxides
The resulting value is NOT normalized (not guaranteed to sum up to 1 or anything special)

This makes NO CHECKS on the presence of fe2/fe3 forms or Fe,
use cleanCats() to be sure.
}
