% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZrSaturation.R
\name{Optical_basicity}
\alias{Optical_basicity}
\title{Optical basicity - the Lambda (Λ) parameter
used in Crisp et al. 2022}
\usage{
Optical_basicity(WR)
}
\arguments{
\item{WR}{Whole rock composition}
}
\description{
Optical basicity - the Lambda (Λ) parameter
used in Crisp et al. 2022
}
