% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{ppxCleanKdTable}
\alias{ppxCleanKdTable}
\title{Prepare a Kd table for use with perplex models:
Create multiple rows for duplicate phases (min_1 min_2 etc)
Add rows with zeros for missing phases
Replace NA's by zeros
NB keep all Kds including phases not in perplex file}
\usage{
ppxCleanKdTable(kd, ppxPhases, interactive = T)
}
\arguments{
\item{kd:}{a kd table (minerals in rows, elements in columns)}

\item{ppxPhases:}{"proper" phases in a perpleX model cube}

\item{interactive:}{Give info messages in pop-up windows}
}
\value{
a matrix of Kds, with names matching ppxPhases
}
\description{
Prepare a Kd table for use with perplex models:
Create multiple rows for duplicate phases (min_1 min_2 etc)
Add rows with zeros for missing phases
Replace NA's by zeros
NB keep all Kds including phases not in perplex file
}
