% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZrSaturation.R
\name{ZrSat_Crisp}
\alias{ZrSat_Crisp}
\title{Zr saturation, Crisp & Berry (2022)
Return (1) the [Zr] of a saturated liquid, at a given T;
(2) the T saturation, at a given [Zr].
If T, or [Zr], is not supplied, the resulting value will be NULL}
\usage{
ZrSat_Crisp(
  WR,
  milli = NULL,
  TK = NULL,
  Pbar = NULL,
  Zr = NULL,
  H2O = NULL,
  ...
)
}
\arguments{
\item{WR}{a data.frame or a matrix containing major element wt%
(it should have only numeric values)}

\item{H2O}{water content (wt%). Will over-ride any version found in WR.}

\item{...}{not used, will be silently ingored, but allows a consistent UI for all ZrSat_* functions}

\item{TK:}{the temperature (in K)}

\item{Pbar:}{pressure (in bars). Note that Crisp uses GPa in the paper!}

\item{Zr:}{[Zr] concentration (ppm)}
}
\description{
Zr saturation, Crisp & Berry (2022)
Return (1) the [Zr] of a saturated liquid, at a given T;
(2) the T saturation, at a given [Zr].
If T, or [Zr], is not supplied, the resulting value will be NULL
}
