% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZrSaturation.R
\name{zrSaturation.GCDkit}
\alias{zrSaturation.GCDkit}
\title{This is the original GCDkit saturation function, kept for reference}
\usage{
zrSaturation.GCDkit(cats = milli, T = 0, Zr = filterOut(WR, "Zr", 1))
}
\description{
This is the original GCDkit saturation function, kept for reference
}
