% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDmodel.R
\name{correctZrnSat}
\alias{correctZrnSat}
\title{Correction for zircon saturation, Watson & Harrison 1983}
\usage{
correctZrnSat(
  kd,
  c0,
  pm,
  cmins,
  min.props,
  melt.arg = list(),
  dont = character(0),
  SatModel = "Boehnke"
)
}
\arguments{
\item{kd}{Named matrix. Partition coefficients. Must include Zircon (Zrn)}

\item{pm}{Scalar. Degree of partial melting (0-100)}

\item{cmins}{Matrix. Composition of minerals (not used, for consistency)}

\item{min.props}{Named vector. Mineral proportions (normalized to 1)}

\item{melt.arg}{List. Further useful arguments.
Includes here the melt major elements (melt.arg$mjrs),trace (melt.arg$trc), temperature in KELVIN (melt.arg$TT)}

\item{dont}{Character vector. Do not calculate for these elements. Use to avoid overptinting previous saturation calc (i.e. if you use a saturation routine /after/ this one, )}

\item{SatModel}{model to use Currently "Boehnke" (the default; Boehnke et al. 2013) or "WH" (Watson & Harrison 1983)}
}
\value{
list containing the following slots:
\itemize{
  \item \code{c0} Named vector. Composition of the source, as passed to the function
  \item \code{cL} Named vector. Composition of the liquid
  \item \code{cmins} Named matrix. Composition of all individual solid mineral phases
  \item \code{cS} Named vector. Composition of the whole solid
  \item \code{min.props} Named vector. Mineral proportions in the restite (and expressed as (0-1) of the restite)
  \item \code{FF} Real. Melt proportion, 0-100, as passed to the function
  \item \code{kd} Named matrix. The Kd table, here similar to the input (included for consistency with other modelling routines)
  \item \code{DD} Named vector. The D (bulk distribution) values.
  \item \code{dont} Character vector. Elements that should not be touched by further correction routines
  \item \code{tot.mass} Real. The real mass of the system (it increased since we added Zr, so it is now 100% + [Zr])
  \item \code{corr.masses} Named vector. Corrected masses for each mineral, sum = tot.mass (NOT 100)
  \item \code{si.used} Real. Amount of SiO2 used to form ZrSiO4, and thus technically not available to major phases (wt %)
  \item \code{o.used} Real. Amount of O2 used to form ZrSiO4, and thus technically not available to major phases (wt %)
  \item \code{sat} Named vectors. Saturation parameters
}
}
\description{
Correction for zircon saturation, Watson & Harrison 1983
}
