% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZrSaturation.R
\name{M_param}
\alias{M_param}
\title{Calculate the M parameter, as used by various zircon saturation equations}
\usage{
M_param(
  cats,
  Moxides = c("SiO2", "TiO2", "Al2O3", "FeOt", "MnO", "MgO", "CaO", "Na2O", "K2O",
    "P2O5")
)
}
\arguments{
\item{cats}{a data.frame containing the cationic proportions,
probably generated by millications() and cleanCations()}

\item{Moxides}{Oxides used to calculate the M parameter. Their molecular
weight will be taken from petroOxides}
}
\value{
a data frame containing the M parameter for each sample
}
\description{
Calculate the M parameter, as used by various zircon saturation equations
}
