% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZrSaturation.R
\name{ZrSat_Crisp}
\alias{ZrSat_Crisp}
\title{Zr saturation, Crisp & Berry (2022)
Return (1) the [Zr] of a saturated liquid, at a given T;
(2) the T saturation, at a given [Zr].
If T, or [Zr], is not supplied, the resulting value will be NULL}
\usage{
ZrSat_Crisp(
  WR,
  milli = NULL,
  TK = NULL,
  Pbar = NULL,
  Zr = NULL,
  H2O = NULL,
  ...
)
}
\arguments{
\item{WR}{a data.frame, matrix or vector containing major element wt\%}

\item{H2O}{water content (wt\%). Will over-ride any version found in WR.}

\item{...}{not used, will be silently ignored, but allows a consistent UI for all ZrSat_* functions}

\item{TK:}{the temperature (in K)}

\item{Pbar:}{pressure (in bars). Note that Crisp uses GPa in the paper!}

\item{Zr:}{[Zr] concentration (ppm)}
}
\value{
A data.frame, containing Zr.sat, T.Zrsat, lambda and xH2O
}
\description{
Zr saturation, Crisp & Berry (2022)
Return (1) the [Zr] of a saturated liquid, at a given T;
(2) the T saturation, at a given [Zr].
If T, or [Zr], is not supplied, the resulting value will be NULL
}
\details{
Calculation done using Crisp and Berry equation. The lambda parameter is calculated
from ancillary function Optical_basicity()
If a T is supplied, the [Zr] saturation for this T is calculated.
If a [Zr] is supplied, either explicitly (using parameter Zr= ...) or
implicitly (in the WR table), a T saturation is calculated.
If both are supplied, both are calculated!
If H2O is explicitly supplied, it is used. Otherwise, the value(s) found in column
H2O of WR. If H2O is not supplied it is assumed to be 0.
Pressure (argument Pbar) is, on the other hand, mandatory and has no sensible defaults...
}
