% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZrSaturation.R
\name{ZrSat_WH}
\alias{ZrSat_WH}
\title{Zr saturation, Watson & Harrison (1983)
Return (1) the [Zr] of a saturated liquid, at a given T;
(2) the T saturation, at a given [Zr].
If T, or [Zr], is not supplied, the resulting value will be NULL}
\usage{
ZrSat_WH(WR, milli = NULL, TK = NULL, Zr = NULL, ...)
}
\arguments{
\item{WR}{a data.frame, matrix or vector containing major element wt\%}

\item{milli}{a data.frame containing millications.
If supplied, it will be used and WR ignored.}

\item{...}{not used, will be silently ignored, but allows a consistent UI for all ZrSat_* functions}

\item{TK:}{the temperature (in K)}

\item{Zr:}{[Zr] concentration (ppm)}
}
\value{
A data.frame, containing Zr.sat, T.Zrsat and M
}
\description{
Zr saturation, Watson & Harrison (1983)
Return (1) the [Zr] of a saturated liquid, at a given T;
(2) the T saturation, at a given [Zr].
If T, or [Zr], is not supplied, the resulting value will be NULL
}
\details{
Calculation done using Watson & Harrison equation. The M parameter is calculated
from ancillary function M_param()
If a T is supplied, the [Zr] saturation for this T is calculated.
If a [Zr] is supplied, either explicitly (using parameter Zr= ...) or
implicitly (in the WR table), a T saturation is calculated.
If both are supplied, both are calculated!
}
