% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZrSaturation.R
\name{ZrSat_WH}
\alias{ZrSat_WH}
\title{Zr saturation, Watson & Harrison (1983)
Return (1) the [Zr] of a saturated liquid, at a given T;
(2) the T saturation, at a given [Zr].
If T, or [Zr], is not supplied, the resulting value will be NULL}
\usage{
ZrSat_WH(WR, TK = NULL, Zr = NULL)
}
\arguments{
\item{WR}{a data.frame or a matrix containing major element wt%
(it should have only numeric values)}

\item{TK:}{the temperature (in K)}

\item{Zr:}{[Zr] concentration (ppm)}
}
\description{
Zr saturation, Watson & Harrison (1983)
Return (1) the [Zr] of a saturated liquid, at a given T;
(2) the T saturation, at a given [Zr].
If T, or [Zr], is not supplied, the resulting value will be NULL
}
